introduce KernelCommandScheduler
